Spark controllers: SchemaDescription lists for triggers & conditions
